+++ src/application/Apexdocs.ts
@@ -2,7 +2,7 @@ import { pipe } from 'fp-ts/function';
 import * as TE from 'fp-ts/TaskEither';
 import * as E from 'fp-ts/Either';
 
-import markdown, { FileWritingError } from './generators/markdown';
+import markdown from './generators/markdown';
 import openApi from './generators/openapi';
 import changelog from './generators/changelog';
 
@@ -10,13 +10,13 @@ import { ApexFileReader } from './apex-file-reader';
 import { DefaultFileSystem } from './file-system';
 import { Logger } from '#utils/logger';
 import {
-  UnparsedSourceFile,
   UserDefinedChangelogConfig,
   UserDefinedConfig,
   UserDefinedMarkdownConfig,
   UserDefinedOpenApiConfig,
 } from '../core/shared/types';
 import { ReflectionError, ReflectionErrors, HookError } from '../core/errors/errors';
+import { FileWritingError } from './errors';
 
 /**
  * Application entry-point to generate documentation out of Apex source files.
@@ -31,12 +31,12 @@ export class Apexdocs {
     try {
       switch (config.targetGenerator) {
         case 'markdown':
-          return await processMarkdown(config);
+          return (await processMarkdown(config))();
         case 'openapi':
           await processOpenApi(config, logger);
           return E.right('✔️ Documentation generated successfully!');
         case 'changelog':
-          return await processChangeLog(config);
+          return (await processChangeLog(config))();
       }
     } catch (error) {
       return E.left([error]);
@@ -51,7 +51,12 @@ async function processMarkdown(config: UserDefinedMarkdownConfig) {
     config.includeMetadata,
     config.exclude,
   );
-  return generateMarkdownDocumentation(fileBodies, config)();
+
+  return pipe(
+    markdown(fileBodies, config),
+    TE.map(() => '✔️ Documentation generated successfully!'),
+    TE.mapLeft(toErrors),
+  );
 }
 
 async function processOpenApi(config: UserDefinedOpenApiConfig, logger: Logger) {
@@ -62,36 +67,6 @@ async function processOpenApi(config: UserDefinedOpenApiConfig, logger: Logger) 
     config.exclude,
   );
   return openApi(logger, fileBodies, config);
-}
-
-function generateDocs(
-  fn: () => TE.TaskEither<ReflectionErrors | HookError | FileWritingError, unknown>,
-): TE.TaskEither<unknown[], string> {
-  return pipe(
-    fn(),
-    TE.map(() => '✔️ Documentation generated successfully!'), // TODO: Different success message for changelog
-    TE.mapLeft((error) => {
-      if (error._tag === 'HookError') {
-        return ['Error(s) occurred while processing hooks. Please review the following issues:', error.error];
-      }
-
-      if (error._tag === 'FileWritingError') {
-        return ['Error(s) occurred while writing files. Please review the following issues:', error.error];
-      }
-
-      return [
-        'Error(s) occurred while parsing files. Please review the following issues:',
-        ...error.errors.map(formatReflectionError),
-      ];
-    }),
-  );
-}
-
-function generateMarkdownDocumentation(
-  fileBodies: UnparsedSourceFile[],
-  config: UserDefinedMarkdownConfig,
-): TE.TaskEither<unknown[], string> {
-  return generateDocs(() => markdown(fileBodies, config));
 }
 
 async function processChangeLog(config: UserDefinedChangelogConfig) {
@@ -109,7 +84,26 @@ async function processChangeLog(config: UserDefinedChangelogConfig) {
     [],
   );
 
-  return generateDocs(() => changelog(previousVersionFiles, currentVersionFiles, config))();
+  return pipe(
+    changelog(previousVersionFiles, currentVersionFiles, config),
+    TE.map(() => '✔️ Changelog generated successfully!'),
+    TE.mapLeft(toErrors),
+  );
+}
+
+function toErrors(error: ReflectionErrors | HookError | FileWritingError): unknown[] {
+  if (error._tag === 'HookError') {
+    return ['Error(s) occurred while processing hooks. Please review the following issues:', error.error];
+  }
+
+  if (error._tag === 'FileWritingError') {
+    return ['Error(s) occurred while writing files. Please review the following issues:', error.error];
+  }
+
+  return [
+    'Error(s) occurred while parsing files. Please review the following issues:',
+    ...error.errors.map(formatReflectionError),
+  ];
 }
 
 function formatReflectionError(error: ReflectionError) {
