+++ src/application/errors.ts
@@ -1,0 +1,8 @@
+export class FileWritingError {
+  readonly _tag = 'FileWritingError';
+
+  constructor(
+    public message: string,
+    public error: unknown,
+  ) {}
+}
